Updated according to review comments & split up tests between V5 and V7
