+++ elasticsearch/src/main/mima-filters/3.0.0.backwards.excludes
@@ -1,30 +1,1 @@
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.ElasticsearchSourceSettings.this")
-ProblemFilters.exclude[IncompatibleMethTypeProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchSource.create")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchSource.create")
-ProblemFilters.exclude[IncompatibleMethTypeProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchSource.typed")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchSource.typed")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchSink.create")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchFlow.create")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchFlow.createWithPassThrough")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchFlow.createBulk")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchFlow.createWithContext")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchFlow.createWithContext")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchFlow.createBulk")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchFlow.createBulk$default$3")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.javadsl.ElasticsearchFlow.createWithPassThrough")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSource.apply")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSource.apply$default$4")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSource.create")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSource.create$default$4")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSource.typed")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSource.typed$default$4")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSink.create")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSink.create$default$3")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchFlow.checkClient")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchFlow.create")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchFlow.create$default$3")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchFlow.createWithContext")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchFlow.createWithContext$default$3")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchFlow.createBulk")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchFlow.createWithPassThrough")
-ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchFlow.createWithPassThrough$default$3")
+ProblemFilters.exclude[Problem]("akka.stream.alpakka.elasticsearch.*)
