+++ elasticsearch/src/test/scala/docs/scaladsl/ElasticsearchSpecBase.scala
@@ -1,0 +1,29 @@
+/*
+ * Copyright (C) 2016-2020 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package docs.scaladsl
+
+import akka.actor.ActorSystem
+import akka.http.scaladsl.{Http, HttpExt}
+import akka.stream.alpakka.testkit.scaladsl.LogCapturing
+import akka.stream.{ActorMaterializer, Materializer}
+import org.scalatest.concurrent.ScalaFutures
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AnyWordSpec
+import org.scalatest.{BeforeAndAfterAll, Inspectors}
+
+trait ElasticsearchSpecBase
+    extends AnyWordSpec
+    with Matchers
+    with ScalaFutures
+    with Inspectors
+    with LogCapturing
+    with BeforeAndAfterAll {
+
+  //#init-mat
+  implicit val system: ActorSystem = ActorSystem()
+  implicit val materializer: Materializer = ActorMaterializer()
+  //#init-mat
+  implicit val http: HttpExt = Http()
+}
