+++ elasticsearch/src/test/scala/docs/scaladsl/ElasticsearchSpecUtils.scala
@@ -1,0 +1,88 @@
+/*
+ * Copyright (C) 2016-2020 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package docs.scaladsl
+
+import akka.http.scaladsl.HttpExt
+import akka.http.scaladsl.model.Uri.Path
+import akka.http.scaladsl.model.{ContentTypes, HttpMethods, HttpRequest, Uri}
+import akka.stream.Materializer
+import akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSource
+import akka.stream.alpakka.elasticsearch.{
+  ApiVersion,
+  ElasticsearchConnectionSettings,
+  ElasticsearchSourceSettings,
+  EsParams
+}
+import akka.stream.scaladsl.Sink
+import org.scalatest.concurrent.ScalaFutures
+import org.scalatest.wordspec.AnyWordSpec
+
+import scala.collection.immutable
+import scala.concurrent.Future
+
+trait ElasticsearchSpecUtils { this: AnyWordSpec with ScalaFutures =>
+  implicit def materializer: Materializer
+
+  def http: HttpExt
+
+  //#define-class
+  import spray.json._
+  import DefaultJsonProtocol._
+
+  case class Book(title: String)
+
+  implicit val format: JsonFormat[Book] = jsonFormat1(Book)
+  //#define-class
+
+  def register(connectionSettings: ElasticsearchConnectionSettings, indexName: String, title: String): Unit = {
+    val request = HttpRequest(HttpMethods.POST)
+      .withUri(Uri(connectionSettings.baseUrl).withPath(Path(s"/$indexName/_doc")))
+      .withEntity(ContentTypes.`application/json`, s"""{"title": "$title"}""")
+    http.singleRequest(request).futureValue
+  }
+
+  def flushAndRefresh(connectionSettings: ElasticsearchConnectionSettings, indexName: String): Unit = {
+    val flushRequest = HttpRequest(HttpMethods.POST)
+      .withUri(Uri(connectionSettings.baseUrl).withPath(Path(s"/$indexName/_flush")))
+    http.singleRequest(flushRequest).futureValue
+
+    val refreshRequest = HttpRequest(HttpMethods.POST)
+      .withUri(Uri(connectionSettings.baseUrl).withPath(Path(s"/$indexName/_refresh")))
+    http.singleRequest(refreshRequest).futureValue
+  }
+
+  def readTitlesFrom(apiVersion: ApiVersion,
+                     sourceSettings: ElasticsearchSourceSettings,
+                     indexName: String): Future[immutable.Seq[String]] =
+    ElasticsearchSource
+      .typed[Book](
+        constructEsParams(indexName, "_doc", apiVersion),
+        query = """{"match_all": {}}""",
+        settings = sourceSettings
+      )
+      .map { message =>
+        message.source.title
+      }
+      .runWith(Sink.seq)
+
+  def insertTestData(connectionSettings: ElasticsearchConnectionSettings): Unit = {
+    register(connectionSettings, "source", "Akka in Action")
+    register(connectionSettings, "source", "Programming in Scala")
+    register(connectionSettings, "source", "Learning Scala")
+    register(connectionSettings, "source", "Scala for Spark in Production")
+    register(connectionSettings, "source", "Scala Puzzlers")
+    register(connectionSettings, "source", "Effective Akka")
+    register(connectionSettings, "source", "Akka Concurrency")
+    flushAndRefresh(connectionSettings, "source")
+  }
+
+  def constructEsParams(indexName: String, typeName: String, apiVersion: ApiVersion): EsParams = {
+    if (apiVersion == ApiVersion.V5) {
+      EsParams.V5(indexName, typeName)
+    } else {
+      EsParams.V7(indexName)
+    }
+  }
+}
